For the tweakers among us: option to cache solver setup data between solves.
This option retains the matrix, preconditioner and solver objects between solves, so that time is saved by not having to rebuild them every solver call, in particular when using multigrid preconditioners. This obviously comes at the price of memory. Also this option only makes sense for matrices that remain the same between solver calls, i.e. the pressure matrix and the balance pressure matrix (not moving your mesh). The solver context containing this information is stored on the (block_)csr_matrix and gets cleared at deallocate and at every call to zero(). If you have a matrix that is changed between calls, without calling zero(), this option won't work (ignoring all inbetween changes).

Switching on this option in balance-pressure for testing purposes.

Discovered a nasty bug in petsc when using "the compare with exact solution" solver diagnostic, where it clobbered PETSC_NULL_OBJECT in the call to KSPBuildSolution.