Hooked up Spatialindex 1.5 as the default. This reduces the computational cost of initialising RTree from nearly 10 mins down to 0.03seconds when running the gyre benchmark with 10m mesh vertices on 1024 processes.
The root of the problem was the earlier version of the spatialindex library was using tempory files and this was hammering the IO to the extent that we were being hit with this increasing cost as the number of processes was increased.

Many thanks to Xiaohu Guo and James Maddison for their help in nailing this.