Fix for double periodic: the positions derived for the inbetween single periodic mesh needs to be consistent.
Changing make_mesh_periodic to also produce the new positions field on the created periodic mesh, with the usual convention of using the "aliased from position" for the nodes on the periodic boundary. This is much easier inside make_mesh_periodic as an explicity list of aliased and physical nodes is available. This also means a few call to post_process_periodic_mesh can be removed.