Attempt at improving the current coordinate field mess.
Due to a lot of new and interesting ways of deriving meshes from another, 
the question "give me a coordinate for a field on this mesh" has become 
a lot more complex. Some of the complications:
- (some) periodic meshes carry around their own coordinate 
field. This periodic coordinate field is required in some 
places, in some places it shouldn't be used. If it is 
required, the coordinate field isn't always present as meshes derived
from a periodic mesh in general won't have one.
- we allow multiple external meshes, but there is only one 
CoordinateMesh. How do we know whether the "Coordinate" field on the 
"CoordinateMesh" is appropriate for a given mesh?
- the fact that a mesh has its own coordinate field (MeshNameCoordinate)
doesn't necessarily mean it should be used. For instance currently 
in a p1p1 setup with a p2 coordinate field, the .stat file still wrongly
uses the p1 coordinate.
- mesh extrusion involves a couple of associated coordinate 
fields. People may be tempted to define fields on the horizontal 
mesh (there are other issues here as well), in this case again a suitable
coordinate field has to be found.

To simplify matters I've come up with 3 distinct cases, that will 
be served by 3 different routines:

1) get_nodal_coordinate_field(state, mesh)
This will always return a positions field on the given mesh even if the
mesh is periodic. In the periodic case it will adhere to the convention 
that nodes on the periodic boundary will get the original position of the
aliased nodes.
2) get_coordinate_field(state, mesh)
This will return a positions field that in each element has the 
same shape as the given mesh. It will never return a periodic coordinate
field and also in the discontinous case the position field may be on a
different mesh (but of the same order).
3) get_diagnostic_coordinate_field(state, mesh)
In most cases this will just return "Coordinate" and is therefore 
the appropriate choice for any fem integration. Only for meshes 
with exclude_from_mesh_adaptivity option (see below) and the 
horizontal mesh in mesh extrusion a different position field 
will be returned. This should for instance be used in the .stat 
file integrations.

get_coordinate_field() is so named because it (almost) has the 
same functionality as the current version. get_nodal_coordinate_field()
and get_diagnostic_coordinate_field() are new. All current calls to
get_coordinate_field() and get_external_coordinate_field() need to be
reconsidered and replaced with an appropriate choice from the 3. None 
of the calls have been replaced yet however. As before the 3 functions 
always return a new reference to a coordinate field, so it
has to be deallocated after.

To simplify the logic for external meshes the following new rule has 
been introduced:
Only one external mesh is allowed without the exclude_from_mesh_adaptivity
option, all other external meshes and their derived meshes need to have 
the exclude_from_mesh_adaptivity option set. All meshes associated with the
CoordinateMesh can therefore easily be recognised (by not having 
the exclude_from_mesh_adaptivity option).

Please let me know if I've missed out any foreseen cases, or if anything 
is unclear.