Adding support for preserving region ids to 1d adaptivity.
This allows bathymetry to be preserved in a crude way (i.e. by specifying region ids and selecting to preserve them) during 1+1d adaptivity.  This could be automated by interleaving coplanar region ids from the bottom of the mesh on the top mesh being adapted.  Probably wouldn't work very well with complex bathymetries (as it won't be able to remove resolution) but seems appropriate for simpler lab based set ups.  Theoretically also already possible with 2+1d adaptivity but untested.

Adding a test for the 1+1d case.  Mostly just checks that the volume of the domain remains constant and that the run finishes (having passed through several interpolations using a supermesh which requires the domain boundaries to remain the same).  Getting this test to pass with debugging enabled required the increasing of the tolerance in the check of the 1d node ownership case, presumably since some new nodes are now coincident with old element boundaries.

Also, increasing the tolerance on the test_vertical_integration unittest as this was failing locally on the unmodified head of the trunk.