Adding a VerticalBalancePressure field that should solve d^2p_{b}/dz^2 = d(rho g)/dz for a type of balance pressure that's equivalent to just the vertical bits of GeostrophicPressure when ignoring coriolis.  This does not work yet... so do not use (it'll FLAbort if anyone tries)!
Removing a gauss point based hack from control volumes when it needs the boundary normals.  This fixes the original cause of mantis issue 809 though doesn't resolve the underlying issue (that transform_to_physical assumes the number of gauss points is whatever the Coordinate field says).  This should allow multimaterial surface tension to run with higher degrees of quadrature however there seem to be no test cases for this so I can't confirm this.

Fixing a potential bug with the free surface, caching of the solver objects and adaptive timesteps.  Now when CMC is modified by the free surface (because the timestep has changed) the solver cache gets destroyed.

Fixing small bug in ifdef'd out code in Conservative_Interpolation.F90.

Increasing the quadrature degree on all multimaterial adaptive test cases.  I hadn't realised that conservative interpolation didn't use it's own local degree (presumably it doesn't because of mantis issue 809) for all aspects of the interpolation.  Hence 1 degree quadrature (which is valid for fixed mesh p0p1cv simulations) is not valid for adaptive simulations that use galerkin projection (except in the lumped mass case - hence those test cases are untouched).