Fourth of four flux turbine tests unleashed:
This is a 2+1d flux turbine testcase.
It uses penalty flux terms to simulate the turbine. This is a very basic testcase where the turbine is basically a opened sluice gate. The result is compared with another simulation where no turbine is applied. Ideally, the results should be the same. Due to the different meshes, the detectors will have slightly different values.