Adding the possibility of using a big spring reference_node with Velocity.  As this is very wrong for most cases but its prominent presence in the options might confuse people there is a large options check that makes sure you are solving the Stokes equation before it allows you to try using it.  Unfortunately, as with the Pressure reference_node, it still can't check whether its usage is valid for every case of Stokes.  In particular note that it currently assumes that you want to set all components of the Velocity at that node to zero which isn't necessarily true (depending on boundary conditions for each component).  Could be generalised quite easily should the need arise.
Also moving some more prognostic field options from the generic fluidity_options file to the prognostic_fields one.

Adding a relative_to_average option to the vector absolute difference fields (identical to the scalar field equivalent) for cases where the answer is only known up to a constant.  This also assumes you want to do this to all components of the vector field but could be generalised easily again.