Column ids are now transferred when using parallel adaptivity with zoltan.  This allows adaptive parallel 2+1d simulations to run although the load balancing is still entirely done in 2d so you don't end up with a very well balanced 3d mesh.  Adding a test case that tests both the presevation of column and region ids through zoltan.
Note that contrary to what I thought when I started this misadventure element_columns are not necessary for the migration of extruded meshes.  Therefore they are defunct for their original purpose.  However, in the meantime they have been used for other purposes in vertical integration.  As I haven't added their migration to zoltan this new functionality won't work in adaptive parallel simulations.

Fixing a bug in zoltan (and reorder_element_numbering) where the reordering wasn't actually being performed because no halos were present at the time (this is in spite of us passing in an element to node list - it just exited silently).  Even fixing this it turns out that the element to node list we were passing it was incorrect as it still contained degenerate elements and used global rather than universal numbering.  Fixing all three of these issues however doesn't seem to change anything so I guess the ordering was fortunately enforced somehow by the migration.

Other snippets:
 - fixing a bug with parallel metric based extrusion introduced by me in r14029
 - adding diagnostic algorithms for the column and universal column ids
 - allowing vtk to write the column data to file and making this the default for adaptivity debugging output
 - increasing the verbosity required to get notification from find_node_ownership_rtree_no_tolerance because 2+1d adaptivity now interpolates once per column which makes this rather noisy
 - allowing renumber_positions_trailing_rceives to return the permutation map and to reorder the column ids internally
 - fixing a bug with ignoring certain meshes in the collect states by mesh