Start of lock-exchange example.
Still needs some fiddling/decision making and will need to update the manual etc once that is sorted but wanted to get the files committed.

The xml is special so shouldn't be run at all and diamond validation seemed happy with it but if it breaks anything feel free to shout or revert!