Petsc's inodes (automatic detection of block structures) caused failures with "mg" preconditioner. In the presence of inodes petsc's SOR algorithm automatically switches over to a block SOR, but because the blocks are selected randomly, independent of the chosen coarsening strategy, this doesn't provide a suitable multigrid smoothing algorithm.
Unfortunately the choice of using inodes or not has to be made at matrix
allocation. Since at that stage we do not know which solver options are going to
be used (in fact different solver options may be applied to the same matrix),
for the moment I have switched of the use inodes for all matrices. This may
however give a performance loss in solves where SOR is chosen as the (main)
preconditioner. Will add some logic later for the cases where we do know the
solver options at allocation of the petsc matrix.