Further improvements to the buoyancy adjustment by vertical diffusion scheme.
There is now the option 'project_buoyancy_to_continuous_space' which calculates the rate of change of buoyancy in the direction of gravity on the field projected to CG space.  This is one approach to allow the scheme to account for differences with respect to neighbouring elements when buoyancy is a discontinuous field.