Restored adaptivity calls in Shallow_Water. Sorry James, I thought these tests were being done from Simple_Shallow_Water, obviously this capability needs to be preserved until your viva.
Just to clarify, plans for Shallow_Water:
* The code is being rewritten by Jemma and I (mostly Jemma) so that it can 
be solved on an arbitrary 2D surface embedded in 3D. The only two cases that
are ever likely to be used are the sphere (well maybe an oblate spheroid if 
someone gets really fancy) and a plane. We will aim to preserve all currently
running planar test cases in the new geometry.
* We are currently at the stage of getting the linear part over to the 
arbitrary 2D surface.
* Next step will be implementing the method of Bernard, Remacle et al for
the nonlinear momentum advection.