- Merging branch revision r15499 into trunk.
Summary of branch changes:

------------------------------------------------------------------------
r15499 | ctj10 | 2011-02-12 03:46:32 +0000 (Sat, 12 Feb 2011) | 5 lines

- Adding a new multi-phase test case called mphase_gravity. Tagging this as 'special' for now since nothing works.

The test case considers a square 2D domain filled with two incompressible fluids starting with a zero velocity ﬁeld. One liquid is denser than the other (\rho_1 = 1.0, \rho_2 = 1.5) and the volume fraction of each phase is initially set to 0.5 everywhere. The heavier liquid should sink beneath the lighter one as time progresses due to the density difference, causing the volume fraction of the heavier liquid to tend to 1.0 and 0.0 in the lower and upper half of the box respectively.