Change the radiation model such that the function space option definitions are now within each /neutral_particle element within the /radiation element rather than the /geometry element.
This prevents non radiation users seeing anything radiation.

The radiation function spaces within the options tree are now templates, 
which are then used to populate the required meshes in an added procedure in 
populate state. This keeps the door open to allow in the future also the 
ability to have different meshes for different sets of energy groups 
within the same neutral_particle (thermal particles require more 
resolution than fast).

This required breaking insert_derived_meshes in to two to create 
insert_derived_mesh, which takes the mesh name and option path as dummy 
variables so as this can be used by the radiation routine also.

Change the radiation tests as needed.