Some tidying of the shallow water code - in particular, the embedded manifold mappings are no longer required (they can be included as part of the python function used to set initial conditions, for example). This means that there is now only one Velocty and Coordinate field within the shallow water code, which reduces confustion considerably!
In order to make this work, we need to be able to read in 3d coordinate fields but setup 2d vector fields. This is achieved by changing the dimension in the header of the .node file to mean the dimension of the mesh rather than the geometry dimension (which can be accessed from the options tree).

The interval script now takes the option --3d so that it can write out 3d coordinate files.

This commit will break the shallow_water_mms tests because I'm removing the meshes committed to svn and the Makefile provided actually generates different meshes. Simon is aware of this and will fix...