Remove the appended word 'Template' to the necessary radiation fields in the schema options and correspondingly in the radiation populate state - looks confusing
Change the name of fission and power fields of different components 
to be more informative in the schema and populate radiation

Stick to solving for one energy group at a time rather than sets of them.

Change the calculation of the eigenvalue to be a spatial integral within 
a energy integral rather than the opposite. This makes it generic to the 
function space within each energy group

Change the assemble of the within group equation such that this is also 
indifferent to the function space of other groups - although it assumes 
the same coordinate mesh for the scatter and production sources but this can
be made generic via interpolation later

Change radiation tests and Makefile as needed

Remove an unecessary procedure in Radiation_Extract_Flux_Field