Again make the radiation model read in of the .radmats a bit more flexible via allowing any case for characters of the keywords
Include some unit tests and also change a test case to test this

This used the fortran functions iachar and achar and assumed a 
decimal value returned. Is this generic enough? I am not sure.

The procedure is 'make_character_upper_case' in 
radiation/Radiation_Materials_Read_Format_Radmats_Base.F90

feedback welcomed on this point