Fill in the functionality to normalise the particle flux that is done after a eigenvalue solve
This requires calculating certain reaction rates over the domain which is
a functionality added in Radiation_Reaction_Rate.F90

This it self requires integrating each energy dof spatially with a 
corresponding material coeff which is generically included in 
Diagnostic_Integrate_Fields.F90 which will integrate the product 
of any number of scalar fields either over the whole domain or 
certain required region id's. The scalar fields must share the 
same positions field.

This isnt called Radiation_ as it is not radiation specific and 
probably belongs in a different directory.

This routine was also useful in calculating the eigenvalue which 
is changed to use this (this integrates the product of 4 scalar fields)

Also there appeared to be unecessary use module statements in certain 
radiation routines (copied around) so these are removed