Get rid of ewrite from Diagnostic_Integrate_Fields.F90
Change the radiation schema and code to do better the scatter energy 
dof iterations via a specification of the direction/s to sweep and solve 
the them. 

No matter what is chosen the first iteration is from highest to lowest 
energy dof to resolve the down scatter.

The possibilies are then HighToLow, LowToHigh, HighToLowToHigh and 
LowToHighToLow. Although LowToHigh seems to be the most useful as 
it aligns with the upscatter physics. 

This doesnt change the answer but 
can reduce the number of group solves needed (by 4 in the test case)
hence providing a speed up. 

change the 6 energy grp test case to test each of the four possibilities
above