Add to assemble/Advection_Diffusion_CG.F90 robin boundary conditions.
Include in the radiation schema fields for the diffusivity and absorption associated with 
each angular moment set of each group set. These options are included for documentation 
as they are populated into particle state and as assemble/Advection_Diffusion_CG.F90 needed the diffusivity 
option path to function - so it may as well come from an actual flml

This included small changes to the radiation populate state for this.

Also change the radiation schema to have the necessary options for solving a diffusion problem 
using assemble/Advection_Diffusion_CG.F90 for its equation type AdvectionDiffusion

This means that for radiation eigenvalue problems the theta term must be set to 1.0 and the 
mass terms removed

change the radiation test cases flml's as needed (hopefully) and have currently only one of them 
use assemble/Advection_Diffusion_CG.F90 rather than the radiation version

Only one as some of the others dont give the right answer yet