Merging branch revisions r15868 and r15869 into trunk.
Summary of changes:
-------------------

------------------------------------------------------------------------
r15869 | ctj10 | 2011-03-22 05:50:05 +0000 (Tue, 22 Mar 2011) | 3 lines

Correcting file names in the mphase_rising_body test's XML file.


------------------------------------------------------------------------
r15868 | ctj10 | 2011-03-22 05:44:27 +0000 (Tue, 22 Mar 2011) | 8 lines

Added a new (short) multiphase test case called 'mphase_rising_body'. This is similar to the mphase_gravity test case committed a while back. Still not quite there though.

The test case considers a square 2D domain filled with two incompressible fluids starting with a zero velocity ﬁeld. One fluid is denser than the other (\rho_1 = 1.0, \rho_2 = 0.9) and the volume fraction of the lighter fluid is set to 0.999 in the bottom tenth of the domain. The lighter body of fluid should rise up through the denser one as time progresses due to the density difference, causing the volume fraction of the lighter fluid to tend to (almost) 1.0 in the upper tenth of the domain, and 0.0 in the rest of the domain.

Video here for the interested reader: http://amcg.ese.ic.ac.uk/~ctj10/movies/mphase_rising_body.avi