More changes to the copy routine:
The schema has been changed to allow us to
use an overlapping option, which gives the right
number for u_nloc and therefore for cv_nonods

I've also added a couple of options under the new
component option

And now we have a galerkin projection of the
PhaseVolumeFraction field output in the vtu. This
could be used for the test cases, although it suffers
from the oscillations we saw before in Chris's
figures, so it might be better to use a functional
form of the analytical solution where possible, and
to compare this to the cv representation of PVF.

The only thing remaining in the copy routine now is
dealing with the component boundary conditions but
I'm too knackered to do that now.