Reorganise quadrature and shape function tests in femtools.
Test routines for quadrature and shape functions are moved to their
corresponding tests in the femtools/tests folder instead of just
calling the test routine in the module.

New modules quadrature_test and shape_functions_test are created
containting all infrastructure related only to the unit tests but not
the actual quadrature and shape_functions modules.

This removes dependencies on test related modules (unittest_tools, spud)
from quadrature and shape_functions. The type quadrature_template in
quadrature had to be made public.