First steps of the new implementation of the user-defined Random walk python extension, since rev 3486 was a bad approach.
Rather that keeping a c_ptr in Fortran, we evaluate the outer Python code and store the local namespace in a global Python dictionary under a given key (done). This way we can access this local context later, and have a nicely generic way of doing Python state with a val function (still needs to be done).