Adding the random_walk_agents_adapt testcase.
This is a copy of the detectors_parallel_adaptivity test case, but with our 3 different sets of agents added. Compared to random_walk_agents test this should do the same, but in 2d and with 4 adapts. And these are currently killing the test, since Zoltan can't handle multiple detector lists yet.