Making Zoltan handle multiple detector lists (in theory).
For this we introduce a function in detector_tools that tells us the buffer size for a detector, given whether we're advecting or adapting. With this we can easily send the id of the parent list with the detector, so Zoltan only needs to do its thing over all registered lists, since the sendlist buffering takes care of everything else.