Initial steps in removing the specifically 1d code from the prototype:
In Copy_outof_into_state:
Sorting out velocity source term to take multi-D input
Generalising the gravity term in the velocity source, so
that the magnitude is given by gravity_magnitude in the schema
and the direction by GravityDirection. The gravity_direction vector
is normalised before the gravity contribution to the vel source is
added.
Some other small changes to sort out little corners of the options io.

Elsewhere:
Making u_source have dimension (ndim * nphases * u_nonods) so we can
read in contributions from a multi-D source. This is actually the shape
implied by the code which actually uses it in multi_dyncore_dg.