- Implemented the fluid-particle drag term in Multiphase.F90. See the manual for instructions on how to use.
Note that the assembly of this term happens *outside* of construct_momentum_cg/dg, and adds the contribution onto mom_rhs. This is to avoid significantly cluttering up the code. Otherwise, the whole state array would be passed into Momentum_CG/DG.F90 and multiple fields from both the fluid and particle phase(s) would need to be passed around.