Implementing a reflecting boundary option for detectors leaving the domain. If this is chosen, instead of turning the detector static, we reflect it on the face through which it leaves the domain.
The geometry of calculating the reflected point seems to be right, but due to the set_stage routine recalculating update_vector for each stage and therefore throwing away the old update_vector, we loose this information and get errors. So, restructuring the inner loops and the set_stage, move_detectors routines is required to store this sufficiently and make all the components play nice with each other.