Commit headline: geostrophic steady state tests now pass for BDM1 and BDFM1. What a lot of work for nothing moving!
The geostrophic steady state is obtained in the following way: the
user provides a streamfunction defined in the correct space (a CG
space with image of the curl in the div-conforming velocity space
e.g. P2 for BDM1 and P2+cubic bubble for BDFM1 although I only did P2
as it doesn't really change much) and selects "balanced" as the
initial condition option under velocity. The velocity is then set from
the curl of the streamfunction (which can be done in local coordinates
because of the commuting pullback and exterior derivative operations).
The pressure is set by projecting f/g*streamfunction into the pressure
space, this works since

<w,fu^\perp> = -<w,f\nabla\psi> = <div w, psi> + vanishing surface term,
which can be balanced against <div w, g D>, and div w is in the pressure space. 
The integration by parts is permitted since w has continuous normal component
and \psi is fully continuous.

All of this applies to the constant f case only, for variable f we get
Rossby waves.

A check is performed in which the obtained Coriolis and pressure gradient
terms are formed in the DG space, and then the combination is projected into
the div-conforming space and checked to be zero. This will need to be switched
off for variable f. 

div_conforming_projection tests the projection into the div-conforming
space using hybridized_helmholtz_solver with -p. No actual test for that yet, it's a bit 
subtle because one needs to check convergence.