Fix a bug in the prototype where it would either run or fail randomly due to something associated with the main.cpp to fortran link. No idea what was causing it but to fix lets use code that we know works (the fluidity main.cpp copied) and mimic that. This required some changes to the prototype fortran wrapper so as to get the simulation name from spud and not attempt to initialise mpi, petsc and python twice.
Fix another bug associated with the prototype time loop in that rather than using the 
option for maximum number of dumps to keep for the finish time (Eh?) actually use the finish time
option!

Also now only do the prototype dumps (not vtu but the old favourate .d.'s) on the option of 
dump period in time steps.

The tempory prototype petsc linked solver is commented in the code next to the hard wired 
option that has turned it off that if used there is a serious memory leak and memory grows
quit rapidly.

Adjust the BL_1d and BL_1d_Gravity test cases for correct timestepping and io options. Get the
BL_1d_DGPress case running with a check against analytic and also a regression reference. Note 
however that refining the mesh doesnt actually reduce the saturation error (slight increase) 
for this case.

Replace BL_test_gravity_unstable (which never appears to have worked) with BL_1d_GravityUnstable that checks 
against analytic and regression reference.

The four running tests are tagged as short as each now takes about 2mins to complete.