Fix element colouring in assemble_advection_diffusion_cg
The dependency graph for CG assembly is such that if any two elements
share a vertex, they must have a different colour.  With the changes
to make_sparsity_transpose (trunk r3540), if we pass the element
connectivity mesh (p0_mesh) and the CG mesh (t%mesh) to
get_csr_sparsity_secondorder, we get exactly this set of
dependencies.  Fixes race conditions in the OpenMP CG assembly loop.