Found a minimal test problem that fails the trace space test. This test initialises a Q1 field and a P1 trace field from the same Python function and checks that the face values of both fields are the same on each face. This passes for 1x1 and 1x2 meshes, but fails for a 2x2 mesh. In the element in which it fails, the Q1 values are
1. 1.5
0. 0.5

and the P1 trace values are
  0.5 1.5
0.5     1.5
0.0     1.0
  0.0 1.0

So the values have been twisted around some how. On the face in which
it is failing, the Q1 values are:
0.5 1.5
and the P1 trace values are:
1.0 1.5

so it is failing on the face with local coordinate component 1 equal
to 1. I have yet to understand whether this is an issue with the local
numbering, or the global numbering.