Changing the schema structure to introduce the concept of Functional Groups.
These define the variables of the biology agents, according to their type, eg. Diatom.
Each FG consists of multiple stage, eg. Living and Dead, with each stage having a different function to update the agent, potentially different RWs, but the same internal variables.
We create a separate detector/agent array for each stage, so that we can apply the same update function, RW, etc to complete agent arrays. This paves the way for incorporating state changes.