Set up an f-sphere test using icosahedral mesh from John Thuburn in tests/bdfm_sphere.
Since the test uses dimensional units on a sphere with Earth radius,
this reminded me that the Piola transform means that the local
coordinates for the velocity get very big for large edge lengths such
as in this testcase. This causes a lot of problems with round-off and
in particular a balanced initialisation is tricky.

So, for the hybridised solver I decided to rescale the local velocity
components by an edge lengthscale obtained from the square-root of the
element area. Ideally you should rescale normal components by the edge
length but that is not possible unless you have a vector-valued local
basis, which we don't. I made a subroutine to compute these scaling
weights, and rescaled all the appropriate matrices.

I was also reminded that the determinant computed in the divergence test 
for the balanced initialisation needs to be rescaled by detJ.