Momentum_Equation: Don't call profiler_tic/toc on deallocated fields
If have_pressure is .false. then a dummy pressure field is created and
subsequently deallocated.  We can't call profiler_tic/toc(p,
"assembly") if p points to the dummy pressure field after it has been
deallocated, since the option path will be full of rubbish.  So move
the relevant tic/toc calls inside the conditionals that guard the
code.