Multiphase prototype changes.
Add code for the input and output of the FE representation
of key magic named fields that are solved with the 
CV scheme of the prototype. Although I noticed that the 
FE representation of density of all phases appears to always 
be zero (the CV solution array has values).

Add initial code for outputting the projected from 
overlapping to DG velocity field. This doesnt work 
yet as there is no '_ndglno' array for this field
internally to the prototype.

Change some test cases to use some of the new output fields.

Remove some unused code.