Replaced VERSION with PACKAGE_VERSION in Makefile,in to be consistent with autoconf naming.
PACKAGE_VERSION is now only set within the root Makefile so that it can be updated at every build, with or without reconfiguring.

Default hehaviour for development branches and trunk is to set PACKAGE_VERSION to $(echo -n `bzr version-info --custom --template="{branch_nick}.{revno}"`). This is currently slow because it communicates with launchpad. A bug report is being submitted to bzr as the required information is all locally available.