Thanks to Lawrence for spotting this bug.
Where there is a line such as:

 [ -f foo ] && bar

in the makefile, and foo doesn't exist, [ -f foo ] returns an error
which gets passed through to the end of the command and which make
then interprets as a build error, and exists. Adding in a || true at
the end of the line fixes this.