Commit purely addressing documentation, orthogonal to functional codebase.
Adding in a short paragraph to the 'Getting Started' chapter noting that unit
testing is broken under Intel as a result of compiler issues, and that we have
no intention of fixing this at our end. Since Intel seem to have no intention
of fixing it at their end either, users are hereby advised not to make unittest
on an Intel build of Fluidity.