Add the capability when using a CG pressure to test the continuity with the CV dual (instead of the usual CG).
This adds the option to the schema, which comments that this is currently 
only for incompressible flow and will make your pressure matrix non symmetric.

State_Matrices is then changed slightly for getting the divergence matrix 
from state via passing in an optional name argument. If not present 
defaults to normal name.

Momentum_Equation is adjusted to include an option that triggers the desired 
behaviour. For incompressible flow this implies if chosen there is a need 
for two ct_m matrices (the left and right are different). This is allocated, 
grabbed from state, which allocates and inserts if not present, forms 
if then uses it and deallocates.

Include 4 initial test cases for this. 

1/ Sinlge phase checking pressure - gravity balance using a range of element types.
2/ mmaterial checking pressure - gravity balance again using a range of element types.
3/ Single phase checking divergence free (from a CV perspective) velocity in 3d using p0p1
4/ A single phase Stokes flow using p2p1 that checks for convergence using mms. 

These are all duplicates of tests already there. The desire is for a CV tested continuity 
with a momentum element pair that works. Some of the above would not work if pressure in 
the momentum was also CV.

Perhaps a few more tests, perhaps a warnig on the schema that this is new and a little unknown 
as well as entries to manual required.