Remove all the code associated with porous media that is not required, not tested, not used and not documented.
Adjust the porous media options such that permiability is
optional and can also be diagonostic, as can the porosity 
now also.

In populate state no need for porous media option checks and
no need to insert the fields in all states.

Adjust the electr... test cases to use python diagnostics 
to form the darcy absorption term.