Fix a bug for advecting fields with hyperc including porosity in that the CFL number needs to account for it.
Also a small correction in metric advection to always have the 
correct variables for cv mass

Add three p0p1 test cases that advect a tracer. The most complex has
two channels, one with a step change in porosity and adapts the mesh
with metric advection that accounts for porosity.