Add a test case to check the fix for velocity children fields set with python diagnostics to be able to have access to all material_phase states.
The physics of this test case is irrelevant.

The short test has two phases with dummy absorption, source and viscosity fields 
in each. These are then used in the corresponding other phase to set the 
velocity absorption, source and viscosity using the python state, the values 
of which are checked in the xml.