In assemble/Diagnostic_Fields_Matrices.F90 do not initialise cv_mass with =>null() (as pointed out by Stephan).
In femtools/FEFields.F90 use ele_val if the porosity is order 0 then 
fill the rest of the array values from this (as pointed out by Stephan).

Move the porous media options into their own file.

Actually in the schema move the porosity options for scalar fields 
to not be part of spatial discretisation. Also rearrange them 
so as to be less confusing with optional options to change the 
porosity_field_name and theta value.

Thus change how these are used in the CV CFL, metric advection 
and Field_Equations_CV. Coupled CV now doesnt expect all fields 
to include porosity.

Slighty adjust the manual for these changes.

Adjust the test cases for these changes.

Including porosity in CG and DG to follow this...