Need to initialise coriolis parameters outside threaded region
A call to coriolis() is not thread-safe if it first occurs inside a
threaded region, since it will call set_coriolis_parameters on first
use.  If multiple threads call this simultaneously we have a race
condition which can lead to incorrect coriolis parameters being set.
To avoid this problem, explicitly initialise the coriolis parameters
outside threaded element assembly in DG and CG Momentum if OpenMP is
enabled.