Change the BL_1d_A.mpml timestep size back to 0.003125 from 0.1, which was changed for soom very unknown strange reason in r3592.
In the copy out of state routine hard set the pressure BC type for all
phase, which is the same as the first phase settings, rather than access 
an array out of its bounds.

The BL_1d now appears to run the same as r3590, where all the set up to 
run test cases have a ~0.5% change compared to r3589 (the last revision 
to pass them). I noticed for this test case also that the saturation 
for the second phase is slightly above 1.0 being around 1.0003 at one 
point. Phase 1 is fine.

The mp proto tests will still fail...