Add a python diagnostic field to this test case for the sum of the phase volume fractions.
Add tests for the min and max of each phase volume fraction 
and the min, max and integral of the sum of the phase 
volume fractions, many of which have unphysical issues.

Add to the test makefile a rm -rf of velocity_project* files 
and the ocvn file which keeps appearing.