Remove option to switch between local and global matrix assembly
Local assembly is necessary for correctness of OpenMP threaded matrix
assembly.  Since it appears to perform no worse than global assembly
for reasonably sized problems, we can remove the global assembly code
path, reducing code complexity.

In addition, we update the assembly routines for CG and DG momentum
and advection-diffusion to assemble the RHS locally (as well as the
existing local assembly of the LHS matrix).  We add a new optional
argument to PetscNumberingCreateVec and set
VEC_IGNORE_OFF_PROC_ENTRIES if it is true.  Vector assembly now no
longer needs to communicate either.

During creation of PETSc matrices from Fluidity csr matrices, we now
do not have to mask non-local rows with -1 in the CG case, since they
will not be communicated anyway.