Remove the Pressure Neumann boundary condition to replace it with normal_flow boundary conditions for the TotalDarcyVelocity and the individual phase DarcyVelocity vector fields.
If there is a TotalDarcyVelocity normal_flow BC on a boundary element 
and a phase DarcyVelocity normal_flow BC also exists then the code 
will exit as one cannot have both.

This required small changes to preprocessor/Boundary_Conditions_From_Options.F90
to include "normal_flow" as a special case for a vector - only a 
scalar surface field is required not a vector. Also there is a small change 
here to allow diagnostic vector fields to populate and initialise any 
BCs found in its path.

Has been checked for TotalDarcyVelocity, but requires a test case committing.

The "normal_value" of the BC is the velocity normal to the domain boundary 
with the normal pointing outwards.