In the Fluidity single phase Darcy model change the Permeability field to AbsolutePermeability as this is what it is and it syncs it with the Fluidity-DarcyIMPES solver. All the single phase darcy tests have been changed as required, as has the manual. The procedures in Populate state have also been slightly changed as required for this.
Add an optional variable to allocate_and_insert_auxilliary_fields called 
force_prescribed_diagnositc_allocate_old_iterated that will force the 
allocation of new memory for diagnostic and prescribed scalar fields for the 
Old and Iterated versions, rather than aliasing them to the latest field. 
This is required for the Darcy-IMPES solver as there are calculations in 
a timestep that require the latest, Old and Iterated values of diagnostic 
and prescribed fields. Not sure why the main Fluidity isnt set up like this ...

Also in Populate state routine allocate_and_insert_auxilliary_fields insert 
as required tensor fields in a similar way to scalar and vector fields. Again
I am not sure why the main Fluidity isnt set up for this ....

Adjust the adaptivity schema options to break it down into option components 
(not changing any functionality) such as to only pick what is needed for the 
Darcy-IMPES options without copying and pasting. Darcy-IMPES will not support 
adapt at first time step (although this could be done through the main fluidity 
and checkpointed) and also metric advection (for now ...)

Set up the Darcy IMPES code to cache the phase face value (which for now is 
only the saturation) and to set it and use it as requird for a consistent 
formulation.