Add options for gravity and density of each phase.
Remove internal algorithm from options and manually do all
diagnostics. This was required for the Density field as the 
femtools version is expecting Velocity fields etc and kept 
complaining.

Add options for strong and weak BC for pressure and saturation.

Density and weak BC require code to go with the options.

Adjust test cases as required and tidy some module use's and the 
makefile dependencies (albeit mannually - requires make makefiles).