Move the procedure that calls python diagnostics for darcy impes fields to main/Darcy_IMPES.F90
Also, add the call to the darcy impes calc of velocity and cfls to the 
main/Darcy_IMPES.F90 as it has to come after the call above.

These are required because the make process compiles in the order:

assemble/
diagnostics/
main/

Update make file dependencies using make makefiles. Tests still run!