Adjust BC for Vphi, pressure and saturation. Remove BC for darcy velocity and total darcy velocity - they can be applied through combinations of vphi and S.
Add a BC to vphi for no_normal_flow that is also used in the saturation 
assemble. Remove the saturation BC zero_flux as this is not required 
because no_normal_flow has been added.

Add a first attempt at weak pressure BC's, although the term still 
requires adding to the saturation equation for consistency.

Add routines to grab the BC, calculate the characteristic length for 
pressure BC. Readjust the cache face value to be types - one for 
each phase. 

Insist that the relperm, absperm and viscosity are treated explicitly 
in the pressure and saturation solves. The end time step values are still 
used in defining a CFL number - maybe this should be changed ...

Adjust test cases to use more weak BC's. Still got issues on surfaces in 
2d and 3d.