Add code to the first phase pressure solve for the gravity source term. Density is always evaluated at old timestep value with upwind value, which is trivial if it is constant.
The gravity field is forced via the options to be on an element wise mesh, 
as it will make calculating vphi easier! Anyway, 101% of the time it is 
going to be a constant field.

Make sure in the pressure assemble when evaluating the vphi at quadrature 
points to use ele_val_at_quad for the gradient_pressure, rather than 
assuming it is element wise. A few more changes like this may well imply 
the model could work for pressure P2, as well as P1.

Calculate the vphi at end of non linear iteration with old relperm, old 
absperm and old viscosity (will also be old density and old gravity). 
Likewise for the CFL number. Both still use the latest pressure gradient.
This makes it consistent with the actually solution. And it is this vphi 
that is used via a FE interpolation to deduce the upwind direction (the one 
part of the algorithm that requires non linear iteration - ignoring the 
ad hoc weak pressure BC and any subcycling. But only if the direction 
of vphi on the face changes in a timestep.)

Adjust test cases as required which still work!