Add own version of ele_grad_at_quad_scalar until checking with the Fluidity powers that be that this is wise. Reason being dont want to assume that the number of quadrature points is decided by the field%mesh%shape%ngi, as I am passing CV related arrays which are different.
Use the above to evaluate the gradient of the capilliary pressure 
of phases > 1 to add a source term to the global continuity equation 
that solves for the first phase pressure. Requires testing, as far 
as I am aware this is only where it requires adding.

Add more cv_shape's to the darcy_impes_type, although in the end 
they may not all be required but can be purged later when the coding 
is finished.

Add a mesh in the schema for the GradientPressure rather than hard 
setting it to the elementwise mesh.

Adjust test cases as required for last point above.