Correct term for capilliary pressure (used gravity variable by mistake!)
Remove InterstitialVelocityPorosity field to be replaced with 
DarcyVelocityOverSaturation. Within the code replace vphi with 
v_over_s. 

Add extra array of fields to the darcy_impes_type for the 
modified_relative_permeability (plus old_ version).

At start of time step copy above latest to old. Each non linear iteration 
after calculating the relative_permeability fields calculate 
modified_relative_permeability = relative_permeability / saturation.

As this divides by saturation include an option for the minimum value 
to use in this in the diamond schema, which by default if not given 
uses 1.0e-06. This is only for calculating modified_relative_permeability, 
which is then used in place of relative_permeability as required throughout 
the code.

Put the calculation of the subcycle timestep size in its own procedure 
to avoid a code duplication.

Adjust test cases as required.