Re-add back capability for prescribed scalar fields and python diagnostic generic scalar fields.
Only add call to new diagnostic routines, calling the old causes 
isses with the Density fields (which is why it was taken out).

This means all Darcy IMPES non python diagnostic fields have algorithm 
internal added to there options. loading a file with this new schema in diamond 
will warn of many additions of this, but everthing else should be the same.

Add options for adapting and interpolation for all fields.

Adjust test cases as required.

Makefile dependencies requires updating.